remove featured image requirement
in the query there was a meta key which required the post to have a
featured image in order to be returned.  This key prevented the use of
post types which didn't have featured images (like testimonials and
events) from displaying.

While that could always be overridden by the available filter, it should
not be the default behavior. Now if users want to only show items which
have a featured image, they add it back in via the
`arconix_flexslider_loop_args` filter.

See the gist here: https://gist.github.com/j-gardner/10437717

fixes #9